Refactor: ColorBottomSheet 리팩토링 - CustomHandler 컴포넌트 분리 - 컬러바텀시트 함수 훅으로 분리
